Set ttl of duplicate sets using number of seconds.
Previously, we were setting the TTL using a calcuated exact time. This
was causing test suites that lock the time in the past to not be able to
use the DuplicateBlocker. Since we just need to make sure these sets get
cleaned up, we can deal with the imprecise nature of setting ttl using
durations.